Restructure into multiple files.
Try to use btc witnessing code.
